Move beacons tests into beacons dir
